Refactor that separates sidebar and sidebar_toggle
