+++ components/sidebar.py
@@ -20,26 +20,8 @@ class Sidebar:
     OPEN_FILTER = "Öppna filter"
     REQUEST_FILL_FORM = "Var god fyll i relevanta sökfält"
 
-    is_open = True
-
     @staticmethod
     def get_component():
-        component = dbc.Row(
-            children=[
-                Sidebar._sidebar_content(),
-                Sidebar._sidebar_button(),
-            ],
-            id="sidebar",
-            className="p-1",
-            style={
-                "background-color": "#DCEAEF",
-                "height": "inherit",
-            },
-        )
-        return component
-
-    @staticmethod
-    def _sidebar_content():
         component = dbc.Col(
             children=[
                 html.H2(Sidebar.SEARCH_FILTER),
@@ -49,24 +31,7 @@ class Sidebar:
                 Sidebar._filter_form(),
             ],
             id="sidebar_content",
-            className="pl-5",
-            style={
-                "height": "inherit",
-            },
-        )
-        return component
-
-    @staticmethod
-    def _sidebar_button():
-        component = dbc.Button(
-            children=[
-                html.I(
-                    id="sidebar_btn_icon",
-                    className="fas fa-chevron-left",
-                )
-            ],
-            id="btn_sidebar",
-            color="primary",
+            className="col-3",
             style={
                 "height": "inherit",
             },
@@ -95,21 +60,3 @@ class Sidebar:
             },
         )
         return component
-
-    # TODO: Possibly modify such that the class style is not overwritten
-    @staticmethod
-    def add_callback(app):
-        @app.callback(
-            Output(component_id="sidebar_content", component_property="style"),
-            Output(component_id="sidebar_btn_icon", component_property="className"),
-            Input(component_id="btn_sidebar", component_property="n_clicks"),
-            prevent_initial_call=True,
-        )
-        def _toggle_sidebar(n_clicks):
-            Sidebar.is_open = not Sidebar.is_open
-            return (
-                {"display": "block"} if Sidebar.is_open else {"display": "none"},
-                "fas fa-chevron-left" if Sidebar.is_open else "fas fa-chevron-right",
-            )
-
-        return app
